extracting project into seperate component and adding scroll-effect
